sparkq/requirements.txt
# Runtime dependencies for SparkQueue (pinned for reproducibility)
fastapi==0.110.3        # REST API in sparkq/src/api.py
uvicorn[standard]==0.29.0  # ASGI server wrapper in sparkq/src/server.py
typer==0.12.3           # CLI commands in sparkq/src/cli.py
pydantic==2.7.1         # Data models in sparkq/src/models.py and FastAPI schemas
pyyaml==6.0.1           # Config parsing for sparkq.yml, ToolRegistry, ScriptIndex
requests==2.31.0        # HTTP error handling in CLI and integrations
anthropic>=0.7.0        # Claude API for prompt execution in sparkq/queue_runner.py
